==> scripts/3b_phenotype_qc.Rmd <==
---
output: 
  pdf_document:
    latex_engine: pdflatex
title: "Phenotype QC for UKB gene-diet interaction analysis"
---

```{r setup, include=F}
knitr::opts_chunk$set(echo=F, message=F,warning=F, fig.path="../output/figures/")
suppressMessages(silent <- lapply(
  c("knitr", "kableExtra", "tidyverse", "cowplot", "jtools"), 
  library, character.only=T))
```

```{r auxiliary}
INT <- function(x) qnorm((rank(x, na.last="keep") - 0.5) / sum(!is.na(x)))
```

```{r read-phenos}
raw_phenos <- read_csv("../data/processed/ukbb_diet_gwis_phenos_raw.csv")
phenos <- read_csv("../data/processed/ukbb_diet_gwis_phenos.csv")
```

All data shown below are from unrelated, European-ancestry individuals in UK Biobank who have not revoked consent for analysis (N ~ 350k). HbA1c distributions then removed individuals with confirmed diabetes (T1D or T2D based on Florez project "prob_poss_..." definitions) and outliers (outside of 3 IQRs from the 25th/75th percentiles) unless otherwise noted.

```{r distributions}
hba1c_raw_hist <- ggplot(raw_phenos, aes(x=hba1c_1)) + 
  geom_histogram() +
  labs(title="HbA1c (raw)") +
  theme(axis.title=element_blank())
hba1c_hist <- ggplot(phenos, aes(x=hba1c)) +
  geom_histogram() +
  labs(title="HbA1c (no DM, + outlier filter)") +
  theme(axis.title=element_blank())
plot_grid(hba1c_raw_hist, hba1c_hist)

top4_ffq_pc_plts <- lapply(1:4, function(i) {
  ggplot(phenos, aes_string(x=paste0("ffq_PC", i))) +
    geom_histogram() +
    labs(title=paste0("FFQ-PC", i), x="")
})
do.call(plot_grid, top4_ffq_pc_plts)

ggplot(raw_phenos, aes(x=sex)) +
  geom_bar()

ggplot(phenos, aes(x=ffq_PC_ERS10)) +
  geom_histogram() +
  labs(x="Environmental risk score based on 10 FFQ-PCs")

ggplot(phenos, aes(x=age)) +
  geom_histogram() +
  labs(x="Age")

ggplot(raw_phenos, aes(x=fct_rev(assessment_centre))) +
  geom_bar() +
  coord_flip() +
  labs(x="Assessment Centre")

ggplot(raw_phenos, aes(x=fct_rev(birthplace))) +
  geom_bar() +
  coord_flip() +
  labs(x="Birthplace")

ggplot(raw_phenos, aes(x=bmi)) +
  geom_histogram() +
  labs(x="BMI")

ggplot(raw_phenos, aes(x=fct_rev(factor(smk_012)))) +
  geom_bar() +
  coord_flip() +
  labs(x="Smoking (No=0, Previous=1, Current=2)")

ggplot(raw_phenos, aes(x=fct_rev(factor(smk_yn)))) +
  geom_bar() +
  coord_flip() +
  labs(x="Smoking (No=0, Yes=1)")
```

## Exposure and covariate associations with HbA1c

```{r hba1c-associations}
lm_df <- phenos %>%  # Retrieves assessment centre and birthplace with their original factor codings
  left_join(select(raw_phenos, id, assessment_centre, birthplace), by="id")

hba1c_ac_lm <- lm(hba1c ~ assessment_centre, data=lm_df)
plot_summs(hba1c_ac_lm)

hba1c_bp_lm <- lm(hba1c ~ birthplace, data=lm_df)
plot_summs(hba1c_bp_lm)

pc_string <- paste0("PC", 1:10, collapse=" + ")
hba1c_pcs_lm <- lm(as.formula(paste0("hba1c ~ ", pc_string)), data=lm_df)
plot_summs(hba1c_pcs_lm)

hba1c_basic_lm <- lm(hba1c ~ age + sex + age_squared, data=lm_df)
plot_summs(hba1c_basic_lm)

all_covar_string <- paste0(pc_string, " + assessment_centre + birthplace + age + sex + age_squared")
hba1c_all_lm <- lm(as.formula(paste0("hba1c ~ ", all_covar_string)), data=lm_df)
plot_summs(hba1c_all_lm, scale=T)

diet_string <- paste0("ffq_PC", 1:10, collapse=" + ")
all_diet_covar_string <- paste0(diet_string, " + ", pc_string, 
                                " + assessment_centre + birthplace + age + sex + age_squared")
hba1c_diet_all_lm <- lm(as.formula(paste0("hba1c ~ ", all_diet_covar_string)), data=lm_df)
plot_summs(hba1c_diet_all_lm, scale=T,
           coefs=paste0("ffq_PC", 1:10))
```

```{r explore-pc2}
self_report_eth <- read_tsv("../data/raw/Florez_app27892_f.21000.0.0_selfreportethnicity.txt") %>%
  select(id=f.eid, self_report_ethnicity=f.21000.0.0)

ggplot(inner_join(lm_df, self_report_eth, by="id"), 
       aes(x=self_report_ethnicity, y=PC2)) +
  geom_boxplot() +
  theme(axis.text.x=element_text(angle=25)) +
  labs(title="What does PC2 represent?")

# ggplot(lm_df, aes(x=birthplace, y=PC2)) +
#   geom_boxplot() +
#   theme(axis.text.x=element_text(angle=45))
# 
# ggplot(lm_df, aes(x=assessment_centre, y=PC2)) +
#   geom_boxplot() +
#   theme(axis.text.x=element_text(angle=45))
```

## Histograms for the full set of 30 FFQ-PCs

Notes:

* 10 FFQ-PCs emerge with a lambda > 1
* The ffq_PC_ERS10 is a weighted environmental risk score (or, equivalently, a measure of diet-predicted HbA1c based on the first 10 FFQ-PCs)

```{r all-ffq-pc-histograms}
ffq_PCs_to_plot <- c(paste0("ffq_PC", 1:10), "ffq_PC_ERS10")
ffq_pc_plts <- lapply(ffq_PCs_to_plot, function(pc) {
  ggplot(phenos, aes_string(x=pc)) +
    geom_histogram() +
    labs(title=gsub("ffq_PC_", "FFQ-PC ", pc), x="")
})
for (plt in ffq_pc_plts) print(plt)
```

```{r additional-variables, eval=F}
acbpAdj_lm <- lm(hba1c ~ assessment_centre + birthplace, data=raw_phenos,
                 na.action=na.exclude)
phenos_expanded <- phenos %>%
  mutate(ffq_median_PC1_INT=INT(ffq_PC1),
         hba1c_acbpResid_INT=INT(resid(acbpAdj_lm)))
write_csv(phenos_expanded, "../data/processed/ukbb_diet_gwis_phenos_expanded.csv")
```
